Ensure "exclusions" are available in event recurrence meta
Fixes a PHP Warning when the `exclusions` key is missing from recurrence
meta.

It seems that it was at one time possible for recurrence meta
to be stored for an event without exclusions data attached. Now,
the code in Events Calendar Pro that expects this exclusions data
will generate a PHP Warning if it is not there.

We've tried opening a support ticket with Modern Tribe, but the
issue has not been resolved upstream. This should be a good enough
workaround.