Use different means to retrieve submitted Community Event IDs
The previous means are no good if a user isn't logged in. It
doesn't seem that a post ID is provided at all for anonymously
submitted events - but the post title is, so we can use the
`get_page_by_title` function to get the post and grab the ID from
that.